move tokenizer
